Create a folder for run related modules - As a first step move out getSynthetixContracts to module
